refactor: separate home page from the main file
