+++ lib/pages/home_page.dart
@@ -1,0 +1,60 @@
+import 'package:dart_plus_app/widgets/search_bar.dart';
+import 'package:dart_plus_app/widgets/title_section.dart';
+import 'package:dart_plus_app/data/mock/fetch/localdataservice.dart';
+import 'package:dart_plus_app/widgets/grid_view_vertical.dart';
+import 'package:flutter/material.dart';
+
+
+class MyHomePage extends StatefulWidget {
+  const MyHomePage({super.key, required this.title});
+  final String title;
+
+  @override
+  State<MyHomePage> createState() => _MyHomePageState();
+}
+
+class _MyHomePageState extends State<MyHomePage> {
+  late Future<List<dynamic>> mediaItems;
+
+  @override
+  void initState() {
+    super.initState();
+    mediaItems = LocalDataService().fetchData();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        backgroundColor: Theme.of(context).colorScheme.background,
+        title: Image.asset('assets/logo.png'),
+        centerTitle: true,
+        //SizedBox(
+         // child: Image.asset('assets/logo.png', alignment: Alignment.center),
+      ),
+      body: Column(
+        children: [
+          const WidgetSearchBar(),
+          const WidgetTitleSection(title: 'Filmes e Séries Populares'),
+          FutureBuilder<List<dynamic>>(
+            future: mediaItems,
+            builder: (context, snapshot) {
+              if (snapshot.connectionState == ConnectionState.done) {
+                if (snapshot.hasError) {
+                  return Center(child: Text('Erro: ${snapshot.error}'));
+                }
+                if (snapshot.hasData) {
+                  return WidgetGridViewVertical(mediaItems: snapshot.data!);
+                } else {
+                  return const Center(child: Text('Nenhum dado disponível'));
+                }
+              } else {
+                return const Center(child: CircularProgressIndicator());
+              }
+            },
+          ),
+        ],
+      ),
+    );
+  }
+}
