+++ pubspec.yaml
@@ -1,5 +1,5 @@
 name: dart_plus_app
-description: "A new Flutter project."
+description: An example of an entertainment stream app.
 # The following line prevents the package from being accidentally published to
 # pub.dev using `flutter pub publish`. This is preferred for private packages.
 publish_to: "none" # Remove this line if you wish to publish to pub.dev
